llm/requirements.txt
llama-cpp-python
pika
python-dotenv
sqlmodel
sqlalchemy-json
requests
huggingface-hub